+++ noodles-vcf/src/variant/record_buf/convert.rs
@@ -1,6 +1,6 @@
 use std::io;
 
-use super::RecordBuf;
+use super::{RecordBuf, Samples};
 use crate::{Header, variant::Record};
 
 impl RecordBuf {
@@ -23,17 +23,45 @@ impl RecordBuf {
     where
         R: Record + ?Sized,
     {
-        use super::Samples;
+        let mut dst = RecordBuf::default();
+        dst.try_clone_from_variant_record(header, record)?;
+        Ok(dst)
+    }
 
-        let mut record_buf = RecordBuf::default();
-
-        *record_buf.reference_sequence_name_mut() = record.reference_sequence_name(header)?.into();
-        *record_buf.variant_start_mut() = record.variant_start().transpose()?;
-        *record_buf.ids_mut() = record.ids().iter().map(String::from).collect();
+    /// Clones the given record into this record.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_vcf::{self as vcf, variant::RecordBuf};
+    ///
+    /// let mut record_buf = RecordBuf::builder()
+    ///     .set_reference_sequence_name("sq0")
+    ///     .build();
+    ///
+    /// let header = vcf::Header::default();
+    /// let record = vcf::Record::default();
+    ///
+    /// record_buf.try_clone_from_variant_record(&header, &record)?;
+    ///
+    /// assert_eq!(record_buf, RecordBuf::default());
+    /// # Ok::<_, std::io::Error>(())
+    /// ```
+    pub fn try_clone_from_variant_record<R>(
+        &mut self,
+        header: &Header,
+        record: &R,
+    ) -> io::Result<()>
+    where
+        R: Record + ?Sized,
+    {
+        *self.reference_sequence_name_mut() = record.reference_sequence_name(header)?.into();
+        *self.variant_start_mut() = record.variant_start().transpose()?;
+        *self.ids_mut() = record.ids().iter().map(String::from).collect();
 
         let raw_reference_bases: Vec<_> =
             record.reference_bases().iter().collect::<io::Result<_>>()?;
-        *record_buf.reference_bases_mut() = String::from_utf8(raw_reference_bases)
+        *self.reference_bases_mut() = String::from_utf8(raw_reference_bases)
             .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
         let raw_alternate_bases: Vec<_> = record
@@ -41,17 +69,17 @@ impl RecordBuf {
             .iter()
             .map(|result| result.map(String::from))
             .collect::<io::Result<_>>()?;
-        *record_buf.alternate_bases_mut() = raw_alternate_bases.into();
+        *self.alternate_bases_mut() = raw_alternate_bases.into();
 
-        *record_buf.quality_score_mut() = record.quality_score().transpose()?;
+        *self.quality_score_mut() = record.quality_score().transpose()?;
 
-        *record_buf.filters_mut() = record
+        *self.filters_mut() = record
             .filters()
             .iter(header)
             .map(|result| result.map(String::from))
             .collect::<io::Result<_>>()?;
 
-        *record_buf.info_mut() = record
+        *self.info_mut() = record
             .info()
             .iter(header)
             .map(|result| {
@@ -81,8 +109,33 @@ impl RecordBuf {
             })
             .collect::<io::Result<_>>()?;
 
-        *record_buf.samples_mut() = Samples::new(keys, values);
+        *self.samples_mut() = Samples::new(keys, values);
 
-        Ok(record_buf)
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use noodles_core::Position;
+
+    use super::*;
+
+    #[test]
+    fn test_try_clone_from_variant_record() -> io::Result<()> {
+        let header = Header::default();
+
+        let record = RecordBuf::builder()
+            .set_reference_sequence_name("sq0")
+            .set_variant_start(Position::MIN)
+            .set_reference_bases("A")
+            .build();
+
+        let mut dst = RecordBuf::default();
+        dst.try_clone_from_variant_record(&header, &record)?;
+
+        assert_eq!(dst, record);
+
+        Ok(())
     }
 }