+++ noodles-vcf/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * vcf/variant/record_buf/convert: Add
+    `RecordBuf::try_clone_from_variant_record` to clone a
+    `vcf::alignment::Record` into an existing `RecordBuf`.
+
 ## 0.93.0 - 2026-08-21
 
 ### Changed